SKA-863: client server communication (#175)
* Wrap SIL Kit RPC API
* Handle RPC terminals
* Add a FMU RPC Manager class

+++ FmuImporter/FmuImporter/SilKit/SilKitCanManager.cs
@@ -62,7 +62,7 @@ public class SilKitCanManager
     }
   }
 
-  public UInt64 AddCanFrameHandler(uint vRef, long vRefIn, CanFrameHandler handler, byte directionMask)
+  public UInt64 AddCanFrameHandler(uint vRef, uint vRefIn, CanFrameHandler handler, byte directionMask)
   {
     CanControllers.TryGetValue(vRef, out var canController);
     if (canController == null)
